provide more and better configuration options
