+++ ios/RollbarReactNative.h
@@ -9,6 +9,7 @@
 
 + (void)initWithAccessToken:(NSString *)accessToken;
 + (void)initWithAccessToken:(NSString *)accessToken configuration:(RollbarConfiguration *)config;
++ (void)initWithAccessToken:(NSString *)accessToken configuration:(RollbarConfiguration*)configuration enableCrashReporter:(BOOL)enable;
 
 // Old methods for backwards compatability
