+++ lib/Rollbar.js
@@ -1,3 +1,3 @@
-Object.defineProperty(exports,"__esModule",{value:true});exports.Configuration=exports.Client=undefined;var _extends=Object.assign||function(target){for(var i=1;i<arguments.length;i++){var source=arguments[i];for(var key in source){if(Object.prototype.hasOwnProperty.call(source,key)){target[key]=source[key];}}}return target;};var _reactNative=require('react-native');var _rollbar=require('rollbar/src/react-native/rollbar');var _rollbar2=_interopRequireDefault(_rollbar);function _interopRequireDefault(obj){return obj&&obj.__esModule?obj:{default:obj};}function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor)){throw new TypeError("Cannot call a class as a function");}}var NativeClient=_reactNative.NativeModules.RollbarReactNative;var Client=exports.Client=function Client(config){var _this=this;_classCallCheck(this,Client);this.captureUncaughtExceptions=function(){if(ErrorUtils){var previousHandler=ErrorUtils.getGlobalHandler();ErrorUtils.setGlobalHandler(function(error,isFatal){if(_this.config.captureUncaught&&_this.config.shouldSend()){_this.error(error,undefined,function(queued){if(previousHandler){previousHandler(error,isFatal);}});}else if(previousHandler){previousHandler(error,isFatal);}});}};this.captureUnhandledRejections=function(){var tracking=require('promise/setimmediate/rejection-tracking');var client=_this;tracking.enable({allRejections:true,onUnhandled:function onUnhandled(id,error){client.error(error);},onHandled:function onHandled(){}});};this.log=function(obj,extra,callback){if(!_this.config.shouldSend()){if(callback){callback(false);}return;}_this.rollbar.log(obj,extra,callback);};this.debug=function(obj,extra,callback){return _this.rollbar.debug(obj,extra,callback);};this.info=function(obj,extra,callback){return _this.rollbar.info(obj,extra,callback);};this.warning=function(obj,extra,callback){return _this.rollbar.warning(obj,extra,callback);};this.error=function(obj,extra,callback){return _this.rollbar.error(obj,extra,callback);};this.critical=function(obj,extra,callback){return _this.rollbar.critical(obj,extra,callback);};this.setPerson=function(id,name,email){_this.rollbar.setPerson({id:id,name:name,email:email});NativeClient.setPerson({id:id,name:name,email:email});};this.clearPerson=function(){_this.rollbar.clearPerson();NativeClient.clearPerson();};if(config instanceof Configuration){this.config=config;}else{this.config=new Configuration(config);}this.config.setPlatform(_reactNative.Platform.OS);this.rollbar=new _rollbar2.default(this.config.toJSON());if(NativeClient){NativeClient.init(this.config.toJSON());this.captureUncaughtExceptions();if(this.config.captureUnhandledRejections){this.captureUnhandledRejections();}}else{throw new Error('Rollbar: Native client not found. Did you run react-native link?');}};var Configuration=exports.Configuration=function Configuration(accessToken,options){var _this2=this;_classCallCheck(this,Configuration);this.shouldSend=function(){return!_this2.releaseStage||!_this2.enabledReleaseStages||_this2.enabledReleaseStages.includes(_this2.releaseStage);};this.setPlatform=function(platform){if(_this2.platform===undefined){if(platform==='ios'||platform==='android'){_this2.platform=platform;}else{_this2.platform='client';}}};this.toJSON=function(){return{accessToken:_this2.accessToken,endpoint:_this2.endpoint,platform:_this2.platform,logLevel:_this2.logLevel,reportLevel:_this2.reportLevel,payload:_extends({codeBundleId:_this2.codeBundleId,releaseStage:_this2.releaseStage,enabledReleaseStages:_this2.enabledReleaseStages,appVersion:_this2.appVersion},_this2.payload),notifier:{name:'rollbar-react-native',version:_this2.version}};};options=options||{};var pkgData=require('../package.json');this.version=pkgData['version'];this.accessToken=accessToken;this.logLevel=options.logLevel||'debug';this.reportLevel=options.reportLevel||'debug';this.endpoint=options.endpoint||'https://api.rollbar.com/api/1/item/';this.appVersion=options.appVersion||undefined;this.codeBundleId=options.codeBundleId||undefined;this.releaseStage=options.releaseStage||undefined;this.enabledReleaseStages=options.enabledReleaseStages||undefined;this.captureUncaught=options.captureUncaught!==undefined?options.captureUncaught:true;this.captureUnhandledRejections=options.captureUnhandledRejections!==undefined?options.captureUnhandledRejections:!__DEV__;this.payload=options.payload||{};};
+Object.defineProperty(exports,"__esModule",{value:true});exports.Configuration=exports.Client=undefined;var _extends=Object.assign||function(target){for(var i=1;i<arguments.length;i++){var source=arguments[i];for(var key in source){if(Object.prototype.hasOwnProperty.call(source,key)){target[key]=source[key];}}}return target;};var _reactNative=require('react-native');var _rollbar=require('rollbar/src/react-native/rollbar');var _rollbar2=_interopRequireDefault(_rollbar);function _interopRequireDefault(obj){return obj&&obj.__esModule?obj:{default:obj};}function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor)){throw new TypeError("Cannot call a class as a function");}}var NativeClient=_reactNative.NativeModules.RollbarReactNative;var Client=exports.Client=function Client(config){var _this=this;_classCallCheck(this,Client);this.captureUncaughtExceptions=function(){if(ErrorUtils){var previousHandler=ErrorUtils.getGlobalHandler();ErrorUtils.setGlobalHandler(function(error,isFatal){if(_this.config.captureUncaught&&_this.config.shouldSend()){_this.error(error,undefined,function(queued){if(previousHandler){previousHandler(error,isFatal);}});}else if(previousHandler){previousHandler(error,isFatal);}});}};this.captureUnhandledRejections=function(){var tracking=require('promise/setimmediate/rejection-tracking');var client=_this;tracking.enable({allRejections:true,onUnhandled:function onUnhandled(id,error){client.error(error);},onHandled:function onHandled(){}});};this.log=function(obj,extra,callback){if(!_this.config.shouldSend()){if(callback){callback(false);}return;}_this.rollbar.log(obj,extra,callback);};this.debug=function(obj,extra,callback){return _this.rollbar.debug(obj,extra,callback);};this.info=function(obj,extra,callback){return _this.rollbar.info(obj,extra,callback);};this.warning=function(obj,extra,callback){return _this.rollbar.warning(obj,extra,callback);};this.error=function(obj,extra,callback){return _this.rollbar.error(obj,extra,callback);};this.critical=function(obj,extra,callback){return _this.rollbar.critical(obj,extra,callback);};this.setPerson=function(id,name,email){_this.rollbar.setPerson({id:id,name:name,email:email});NativeClient.setPerson({id:id,name:name,email:email});};this.clearPerson=function(){_this.rollbar.clearPerson();NativeClient.clearPerson();};if(config instanceof Configuration){this.config=config;}else{this.config=new Configuration(config);}this.config.setPlatform(_reactNative.Platform.OS);this.rollbar=new _rollbar2.default(this.config.toJSON());if(NativeClient){NativeClient.init(this.config.toJSON());this.captureUncaughtExceptions();if(this.config.captureUnhandledRejections){this.captureUnhandledRejections();}}else{throw new Error('Rollbar: Native client not found. Did you run react-native link?');}};var Configuration=exports.Configuration=function Configuration(accessToken,options){var _this2=this;_classCallCheck(this,Configuration);this.shouldSend=function(){return!_this2.releaseStage||!_this2.enabledReleaseStages||_this2.enabledReleaseStages.includes(_this2.releaseStage);};this.setPlatform=function(platform){if(_this2.platform===undefined){if(platform==='ios'||platform==='android'){_this2.platform=platform;}else{_this2.platform='client';}}};this.toJSON=function(){return{accessToken:_this2.accessToken,endpoint:_this2.endpoint,platform:_this2.platform,logLevel:_this2.logLevel,reportLevel:_this2.reportLevel,enabled:_this2.enabled,verbose:_this2.verbose,transform:_this2.transform,payload:_extends({codeBundleId:_this2.codeBundleId,releaseStage:_this2.releaseStage,enabledReleaseStages:_this2.enabledReleaseStages,appVersion:_this2.appVersion},_this2.payload),notifier:{name:'rollbar-react-native',version:_this2.version}};};options=options||{};var pkgData=require('../package.json');this.version=pkgData['version'];this.accessToken=accessToken;this.logLevel=options.logLevel||'debug';this.reportLevel=options.reportLevel||'debug';this.endpoint=options.endpoint||'https://api.rollbar.com/api/1/item/';this.appVersion=options.appVersion||undefined;this.codeBundleId=options.codeBundleId||undefined;this.releaseStage=options.releaseStage||undefined;this.enabledReleaseStages=options.enabledReleaseStages||undefined;this.captureUncaught=options.captureUncaught!==undefined?options.captureUncaught:true;this.captureUnhandledRejections=options.captureUnhandledRejections!==undefined?options.captureUnhandledRejections:!__DEV__;this.payload=options.payload||{};this.enabled=options.enabled===undefined?true:options.enabled;this.verbose=options.verbose||false;this.transform=options.transform;};
 
 //# sourceMappingURL=Rollbar.js.map
