+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rollbar-react-native",
-  "version": "0.2.0-alpha3",
+  "version": "0.2.0-alpha4",
   "description": "Library for reporting errors to Rollbar from React Native apps",
   "main": "index.js",
   "repository": "https://github.com/rollbar/rollbar-react-native.git",
@@ -17,7 +17,7 @@
   "dependencies": {
     "buffer": "^5.0.7",
     "promise": "^8.0.1",
-    "rollbar": "^2.3.7",
+    "rollbar": "^2.3.9",
     "url": "^0.11.0"
   },
   "devDependencies": {
